Move the logic to the Main controller
