+++ Rocket.Chat/Controllers/Chat/MainChatViewController.swift
@@ -9,7 +9,7 @@
 import UIKit
 import SideMenuController
 
-class MainChatViewController: SideMenuController {
+class MainChatViewController: SideMenuController, SideMenuControllerDelegate {
 
     class func shared() -> MainChatViewController? {
         return UIApplication.shared.windows.first?.rootViewController as? MainChatViewController
@@ -37,8 +37,33 @@ class MainChatViewController: SideMenuController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
+        self.delegate = self
+
         performSegue(withIdentifier: "showCenterController", sender: nil)
         performSegue(withIdentifier: "containSideMenu", sender: nil)
+    }
+
+
+    // MARK: SideMenuControllerDelegate
+
+    func sideMenuControllerWillHide(_ sideMenuController: SideMenuController) {
+        ChatViewController.sharedInstance()?.textView.resignFirstResponder()
+        SubscriptionsViewController.sharedInstance()?.textFieldSearch.resignFirstResponder()
+    }
+
+    func sideMenuControllerDidHide(_ sideMenuController: SideMenuController) {
+        ChatViewController.sharedInstance()?.textView.resignFirstResponder()
+        SubscriptionsViewController.sharedInstance()?.textFieldSearch.resignFirstResponder()
+    }
+
+    func sideMenuControllerDidReveal(_ sideMenuController: SideMenuController) {
+        ChatViewController.sharedInstance()?.textView.resignFirstResponder()
+        SubscriptionsViewController.sharedInstance()?.textFieldSearch.resignFirstResponder()
+    }
+
+    func sideMenuControllerWillReveal(_ sideMenuController: SideMenuController) {
+        ChatViewController.sharedInstance()?.textView.resignFirstResponder()
+        SubscriptionsViewController.sharedInstance()?.textFieldSearch.resignFirstResponder()
     }
 
 }
